fix : refactor paths after restructuring the domain module
